Move `validatePerson` function
`src/models/person.js` and `src/models/methods/person_methods.js`:
* Move `validatePerson` function from `person_methods.js` back to `person.js` to avoid confusion, as it isn't actually a method.

`src/models/staff.js` and `src/models/student.js`:
* Update references to `validatePerson` function.
