+++ src/models/person.js
@@ -1,7 +1,7 @@
 // Person: Base class and methods for all user account types
 
 const {
-  ObjectId, enums, privateAttr: { privateAttr, privateAttrStr },
+  ObjectId, enums, privileges, privateAttr: { privateAttr, privateAttrStr },
 } = require('./base');
 const { Faculty } = require('./faculty');
 const { Department } = require('./department');
@@ -34,8 +34,28 @@ const person = {
   resetOTP: String,
 };
 
+/**
+ * Validations and constraints for user accounts. Enforces some
+ * default values for account creation/updating for compliance
+ * with the business logic, security, or both in some cases.
+ */
+function validatePerson() {
+  // Ensure new person has defaults where applicable
+  if (this.isNew) {
+    [this.status] = statuses;
+    this.password = undefined;
+    this.resetPwd = undefined; this.resetTTL = undefined; this.resetOTP = undefined;
+  } else {
+    // Activate a new user account when the user sets a password
+    this.status = this.password && this.status === statuses[0] ? statuses[1] : this.status;
+  }
+  // Set staff privileges based on assigned role (and prevent manual reassignment)
+  this.privileges = privileges[this.role];
+}
+
 module.exports = {
   person,
+  validatePerson,
   methods,
   privateAttr,
   privateAttrStr,
